@@ -1,15 +1,16 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 30
+    buildToolsVersion "30.0.3"
 
     defaultConfig {
         applicationId "com.ereinecke.location"
         minSdkVersion 17
-        targetSdkVersion 23
-        versionCode 1
-        versionName "1.0"
+        targetSdkVersion 30
+        multiDexEnabled true
+        versionCode 2
+        versionName "1.1"
     }
     buildTypes {
         release {
@@ -17,20 +18,25 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
+    compileOptions {
+        sourceCompatibility 1.8
+        targetCompatibility 1.8
+    }
 }
 
 ext {
-    supportLibVersion = '25.3.1'
-    playServicesVersion = '8.4.0'
+    appCompatVersion = '1.2.0'
+    playServicesVersion = '12.0.1'
 }
 
 dependencies {
-
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-    compile "com.android.support:appcompat-v7:${supportLibVersion}"
-    compile "com.android.support:mediarouter-v7:${supportLibVersion}"
-    compile "com.google.android.gms:play-services:${playServicesVersion}"
+    implementation 'com.android.support:multidex:1.0.3'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    testImplementation 'junit:junit:4.12'
+    implementation "androidx.appcompat:appcompat:${appCompatVersion}"
+    implementation 'androidx.mediarouter:mediarouter:1.2.1'
+    implementation "com.google.android.gms:play-services:${playServicesVersion}"
 }
 
 apply plugin: 'com.google.gms.google-services'
